Refactor to make info_gathering more maintable
