+++ lib/main.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:inkstep/blocs/journey_bloc.dart';
 import 'package:inkstep/ui/pages/journey_page.dart';
+import 'package:inkstep/ui/pages/onboarding.dart';
 
 void main() => runApp(Inkstep());
 
@@ -57,7 +58,7 @@ class InkstepState extends State<Inkstep> {
         ),
       ),
       home: BlocProvider<JourneyBloc>(
-        child: JourneyPage(), //Onboarding(),
+        child: Onboarding(),
         bloc: _journeyBloc,
       ),
     );
