+++ lib/ui/components/logo.dart
@@ -29,7 +29,7 @@ class GrowTransition extends StatelessWidget {
 
 class Logo extends StatefulWidget {
   @override
-  _LogoState createState() => _LogoState(80.0, 90.0);
+  _LogoState createState() => _LogoState(90.0, 90.0);
 }
 
 class _LogoState extends State<Logo> with SingleTickerProviderStateMixin {
